fix(file.sls): squash bug related to init failure
The service definition is created and blocks all future attempts to
accept the license. This is important if a user needs to reset or
re-validate their config by placing a 0 byte file named 'ftr' in the
$SPLUNK_HOME.
This is also important if the version changes in the pillar data as
the accept license step may be needed with a large enough version jump.